refactor(dao): JdbcTemplate을 core 패키지로 이동
